requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
asyncpg>=0.29.0  # PostgreSQL async driver
alembic>=1.13.1  # Database migrations
websockets>=12.0  # WebSocket support
redis>=5.0.1  # Session management
aiohttp>=3.9.3  # Async HTTP client
greenlet>=3.0.3  # Required for async SQLAlchemy

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.5  # Async testing support
httpx>=0.25.1  # For async testing

# Development
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6  # For form data parsing 